Remove qcommon and added q.module.d.ts which declares a module 'q'. Also added a couple jasmine style tests.
